refactor: 💡 change project structure
change project structure(move styled components to separate files, move
routes to Page folder)
