+++ src/pages/authentication/authentication.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-export const Authentication: React.FC = ({ match }: any) => {
+export const authentication: React.FC = ({ match }: any) => {
   return (
     <div>{match.path === "/login" ? "Please Log In" : "Please register"}</div>
   );
